Add files via upload
Programming Problems
1. Problem Description
We want to develop a program which manages the purchasing of products. There are 10 different types of products in a stock. Each product has a unique identification number from 1 to 10. We want to maintain the price of each product type, the identification number of each product type and the number of items of each kind of product that are purchased.  A customer can buy several products at once. For example, you can show purchasing information such as:
•	 1 item from product type 5,
•	 4 items from product 2,
•	 7 items from product type 1. 
Given the above information of a product, the program should perform the following functions.

1. View the unit price of a product
2. Add an item to the information system when a product is purchased
3. Calculate the total price of a product